+++ packages/expo-router/src/fork/createMemoryHistory.ts
@@ -172,10 +172,11 @@ export function createMemoryHistory() {
 
         pending.push({ ref: done, cb: done });
 
-        // If navigation didn't happen within 100ms, assume that it won't happen
+        // If navigation didn't happen within 1000ms, assume that it won't happen
         // This may not be accurate, but hopefully it won't take so much time
         // In Chrome, navigation seems to happen instantly in next microtask
-        // But on Firefox, it seems to take much longer, around 50ms from our testing
+        // But on Firefox, the traversal can take much longer when the main thread
+        // is busy, e.g. it was measured to take up to ~900ms while rendering a new screen
         // We're using a hacky timeout since there doesn't seem to be way to know for sure
         const timer = setTimeout(() => {
           const index = pending.findIndex((it) => it.ref === done);
@@ -184,7 +185,7 @@ export function createMemoryHistory() {
             pending[index]!.cb();
             pending.splice(index, 1);
           }
-        }, 100);
+        }, 1000);
 
         const onPopState = () => {
           const id = window.history.state?.id;