+++ packages/expo-router/CHANGELOG.md
@@ -77,6 +77,7 @@
 
 ### 🐛 Bug fixes
 
+- Prevent a delayed `popstate` on web from resetting navigation state after a programmatic `history.go()`. ([#PRNUMBER](https://github.com/expo/expo/pull/PRNUMBER) by [@ahmdshrif](https://github.com/ahmdshrif))
 - Oder tabs by `.Trigger` order during initial render ([#49848](https://github.com/expo/expo/pull/49848) by [@Ubax](https://github.com/Ubax))
 - Prevent Native Tabs from remounting the focused tab while preloading other tabs after a cold-start deep link. (by [@Ubax](https://github.com/Ubax)) ([#49811](https://github.com/expo/expo/pull/49811) by [@Ubax](https://github.com/Ubax))
 - Re-export the vendored JavaScript stack API from `expo-router/js-stack`. ([#49657](https://github.com/expo/expo/pull/49657) by [@davidmokos](https://github.com/davidmokos))